+++ drone/src/maths.cpp
@@ -13,7 +13,7 @@ float PID::operator()(float value, float desire, float p, float i, float d) {
 		return 0;
 	}
 
-	size_t nowMicros = micros();
+	unsigned long nowMicros = micros();
 	float dt = (float)(nowMicros - m_lastTimeMicros) / 1e6f;
 	m_lastTimeMicros = nowMicros;
